Created static website
Implemented a static website using JavaScript to handle all the database data, that was passed to a JavaScript file, in an array, through a python script. That way it is easier to maintain and to deploy. And updated the CSS files, now it is more responsive. To do now is update the content of the site, make small changes and add info section, and update the readme.